Moved all of hooks from utils into hooks directory.
