+++ src/renderer/hooks/useGuid.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import uuidv4 from './uuidv4';
+import uuidv4 from 'renderer/utils/uuidv4';
 
 export default function useGuid() {
   // eslint-disable-next-line no-unused-vars
